+++ src/main/java/controller/RequestMapping.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2019 LINE Corporation. All rights reserved.
  * LINE Corporation PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
  */
-package webserver;
+package controller;
 
 import request.HttpMethod;
