+++ src/main/java/handler/RequestStrategy.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2019 LINE Corporation. All rights reserved.
  * LINE Corporation PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
  */
-package webserver.handler;
+package handler;
 
 import request.RequestHeader;
 import response.Response;
