+++ src/main/java/handler/StaticMappingHandler.java
@@ -1,4 +1,4 @@
-package webserver.handler;
+package handler;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
