+++ src/test/java/webserver/http/RequestStrategyTest.java
@@ -1,12 +1,9 @@
-/*
- * Copyright (c) 2019 LINE Corporation. All rights reserved.
- * LINE Corporation PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
- */
 package webserver.http;
 
 import org.junit.jupiter.api.Test;
-import request.HttpMethod;
 import request.RequestHeader;
+import response.Response;
+import handler.RequestEngine;
 
 import java.io.BufferedReader;
 import java.io.StringReader;
@@ -14,12 +11,12 @@ import java.io.StringReader;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
- * Created by youngjae.havi on 2019-08-02
+ * Created by youngjae.havi on 2019-08-03
  */
-public class RequestMappingHeaderTest {
+public class RequestStrategyTest {
 
     @Test
-    void read_header() throws Exception {
+    void find_correct_strategy() throws Exception {
         BufferedReader bufferedReader = new BufferedReader(new StringReader("GET /index.html HTTP/1.1\n" +
                 "Host: www.nowhere123.com\n" +
                 "Accept: image/gif, image/jpeg, */*\n" +
@@ -28,7 +25,9 @@ public class RequestMappingHeaderTest {
                 "User-Agent: Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1)"));
         RequestHeader requestHeader = new RequestHeader(bufferedReader);
 
-        assertThat(requestHeader.getRequestLine().getMethod()).isEqualTo(HttpMethod.GET);
-        assertThat(requestHeader.getHost()).isEqualTo(" www.nowhere123.com");
+        RequestEngine requestEngine = new RequestEngine();
+        Response response = requestEngine.run(requestHeader);
+
+        assertThat(response.getBody()).isNotEmpty();
     }
 }
